+++ pybamm/models/base_model.py
@@ -714,9 +714,6 @@ class BaseModel(object):
         "Return default solver based on whether model is ODE model or DAE model"
         if len(self.algebraic) == 0:
             return pybamm.ScipySolver()
-        elif pybamm.have_idaklu() and self.use_jacobian is True:
-            # KLU solver requires jacobian to be provided
-            return pybamm.IDAKLUSolver()
         else:
             return pybamm.CasadiSolver(mode="safe")
 
@@ -740,11 +737,7 @@ def find_symbol_in_dict(dic, name):
 
 
 def find_symbol_in_model(model, name):
-    dics = [
-        model.rhs,
-        model.algebraic,
-        model.variables,
-    ]
+    dics = [model.rhs, model.algebraic, model.variables]
     for dic in dics:
         dic_return = find_symbol_in_dict(dic, name)
         if dic_return: