+++ tests/integration/test_models/test_full_battery_models/test_lead_acid/test_full.py
@@ -124,7 +124,6 @@ class TestLeadAcidFullSurfaceForm(unittest.TestCase):
 
 if __name__ == "__main__":
     print("Add -v for more debug output")
-    # pybamm.set_logging_level("DEBUG")
     import sys
 
     if "-v" in sys.argv: